+++ release_notes.md
@@ -1,6 +1,7 @@
 ###In Development
   - [#194](https://github.com/MehdiK/Humanizer/pull/194): Added pt-BR NumberToWords localisation
   - [#147](https://github.com/MehdiK/Humanizer/pull/147): Added Russian translation for ToWords
+  - [#190](https://github.com/MehdiK/Humanizer/pull/190): Added French translation for ToWords and ToOrdinalWords
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.20.15...master)