+++ release_notes.md
@@ -4,6 +4,7 @@
   - [#190](https://github.com/MehdiK/Humanizer/pull/190): Added French translation for ToWords and ToOrdinalWords
   - [#179](https://github.com/MehdiK/Humanizer/pull/179): Adding Hungarian localisation
   - [#181](https://github.com/Mehdik/Humanizer/pull/181): Added Bulgarian localization, date and timespan tests
+  - [#148](https://github.com/Mehdik/Humanizer/pull/148): Adding Hebrew localization for date and timespan
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.20.15...master)